feat: Add functions to create, list, open, and stop profiles
- Added a new function `create_profile_details` to create profile details with proxy information.
- Added a new function `list_browsers` to list all available browsers using an API call.
- Added a new function `close_browsers` to gracefully close opened browsers.
- Added a new function `delete_profiles` to delete profiles using an API call.
- Added a new function `open_profile` to open a specific profile using Selenium.
- Added a new function `upload_extension` to upload an extension using an API call.
- Renamed the file `browser_manager.py` to `list_and_open_profiles.py`.
- Added a new function `update_profile` to update profile details using an API call.
- Added a new function `update_profile_status` to update the status of a profile using an API call.
- Added a new function `start_profiles` to create and open multiple profiles.
- Added a new function `stop_profiles` to close browsers and delete profiles.

These changes provide the necessary functionality to manage profiles and perform actions like creating, listing, opening, and stopping profiles.
