+++ core/browser/create_profiles_in_bulk.py
@@ -62,36 +62,3 @@ def create_profiles_in_bulk(profiles, auth_token):
     else:
         print(f"Error creating profiles: {response.text}")
         return None
-
-def update_profile(profile_id, updated_data, auth_token):
-    url = f"https://dolphin-anty-api.com/browser_profiles/{profile_id}"
-    headers = {
-        'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + auth_token
-    }
-    payload = json.dumps(updated_data)
-
-    response = requests.put(url, headers=headers, data=payload)
-    if response.status_code == 200:
-        print(f"Profile {profile_id} updated successfully.")
-        return response.json()
-    else:
-        print(f"Error updating profile {profile_id}: {response.text}")
-        return None
-
-def upload_extension(file_path, auth_token):
-    url = "https://dolphin-anty-api.com/extensions/upload-zipped"
-    headers = {
-        'Authorization': 'Bearer ' + auth_token,
-        'Content-Type': 'application/json'
-    }
-    files = {
-        'file': open(file_path, 'rb')
-    }
-    response = requests.post(url, headers=headers, files=files)
-    if response.status_code == 200:
-        print("Extension uploaded successfully.")
-        return response.json()
-    else:
-        print(f"Error uploading extension: {response.text}")
-        return None
