+++ core/profile/create_profile_details.py
@@ -1,0 +1,10 @@
+def create_profile_details(proxy_details, index):
+    return {
+        "name": f"browser_{index+1}",
+        "tags": ["rise_extension-profiles"],
+        "mainWebsite": "",
+        "notes": {
+            "content": None
+        },
+        "proxy": proxy_details
+    }
